+++ b2sdk/b2http.py
@@ -23,7 +23,8 @@ from typing import Any, Dict, Optional
 
 from .exception import (
     B2Error, B2RequestTimeoutDuringUpload, BadDateFormat, BrokenPipe, B2ConnectionError,
-    B2RequestTimeout, ClockSkew, ConnectionReset, interpret_b2_error, UnknownError, UnknownHost
+    B2RequestTimeout, ClockSkew, ConnectionReset, interpret_b2_error, UnknownError, UnknownHost,
+    InvalidJsonResponse
 )
 from .api_config import B2HttpApiConfig, DEFAULT_HTTP_API_CONFIG
 from .requests import NotDecompressingResponse
@@ -165,6 +166,7 @@ class B2Http:
     TRY_COUNT_DOWNLOAD = 20
     TRY_COUNT_HEAD = 5
     TRY_COUNT_OTHER = 5
+    NON_JSON_BYTES_COUNT = 200
 
     def __init__(self, api_config: B2HttpApiConfig = DEFAULT_HTTP_API_CONFIG):
         """
@@ -384,6 +386,7 @@ class B2Http:
 
         :param dict post_params: request parameters
         """
+        response = None
         try:
             response = fcn()
             if response.status_code not in [200, 206]:
@@ -427,6 +430,15 @@ class B2Http:
 
         except requests.Timeout as e:
             raise B2RequestTimeout(str(e))
+
+        except json.JSONDecodeError:
+            # When the user points to an S3 endpoint, he won't receive the JSON error
+            # he expects. In that case, we can provide at least a hint of "what happened".
+            # Original JSONDecodeError message, while useful, is not human-readable enough.
+            if response is None:
+                raise RuntimeError('Got JSON error without a response.')
+            raw_content = response.content[:cls.NON_JSON_BYTES_COUNT]
+            raise InvalidJsonResponse(raw_content)
 
         except Exception as e:
             text = repr(e)