+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Authorizing a key for a single bucket ensures that this bucket is cached
 * `Bucket.ls` operation supports wildcard matching strings
 * Documentation for AbstractUploadSource and its children
+* `InvalidJsonResponse` when the received error is not a proper JSON document
 
 ### Fixed
 * Removed information about replication being in closed beta